requirements.txt
astral==1.6.1
click==7.0
paho-mqtt==1.4.0
schedule==0.5.0
